test: CarsTest, CarTest 분리
